+++ structs.go
@@ -1,0 +1,7 @@
+package main
+
+// Action ...
+type Action struct {
+	Payload map[string]interface{} `json:"payload"`
+	Type    string                 `json:"type"`
+}
